requirements.txt
langgraph==0.3.20
langchain-openai==0.3.10
langchain==0.3.23
langchain_community==0.3.21
pydantic-settings==2.8.1
fastapi==0.115.12
python-dotenv==1.1.0
uvicorn==0.34.0
pyjwt==2.10.1
python-jose==3.4.0
sqlalchemy==2.0.39
passlib==1.7.4
psycopg_binary==3.2.6
psycopg2_binary==2.9.10
psycopg-pool==3.2.6
langgraph-checkpoint-postgres==2.0.19
python-multipart==0.0.20
python-docx==1.1.2
openai-whisper==20240930
soundfile==0.13.1
librosa==0.11.0
websockets==15.0.1
sounddevice==0.5.1
pyttsx3==2.98
streamlit